chore: migrate to newer bundler setup
