+++ .gitignore
@@ -34,6 +34,7 @@ stats-hydration.json
 stats-react.json
 stats.html
 .vscode/settings.json
+.idea/
 
 *.log
 .DS_Store
