+++ nx.json
@@ -1,0 +1,74 @@
+{
+  "$schema": "./node_modules/nx/schemas/nx-schema.json",
+  "affected": {
+    "defaultBase": "main"
+  },
+  "tasksRunnerOptions": {
+    "default": {
+      "runner": "nx-cloud",
+      "options": {
+        "cacheableOperations": [
+          "test:lib",
+          "test:eslint",
+          "test:types",
+          "test:build",
+          "build"
+        ],
+        "parallel": 5,
+        "accessToken": "OTI3Y2U3NGQtYzQ3ZC00ZmE3LWJjZWQtYTYxOTEyNmNiN2IyfHJlYWQtb25seQ=="
+      }
+    }
+  },
+  "defaultBase": "main",
+  "pluginsConfig": {
+    "@nrwl/js": {
+      "analyzeSourceFiles": false
+    }
+  },
+  "namedInputs": {
+    "sharedGlobals": [
+      "{workspaceRoot}/.browserslistrc",
+      "{workspaceRoot}/.eslintrc.cjs",
+      "{workspaceRoot}/babel.config.cjs",
+      "{workspaceRoot}/package.json",
+      "{workspaceRoot}/scripts/getRollupConfig.js",
+      "{workspaceRoot}/tsconfig.json"
+    ],
+    "default": [
+      "sharedGlobals",
+      "{projectRoot}/**/*",
+      "!{projectRoot}/**/*.md"
+    ],
+    "public": [
+      "default",
+      "{projectRoot}/build",
+      "{projectRoot}/dist",
+      "!{projectRoot}/.eslintrc.cjs",
+      "!{projectRoot}/tsconfig.eslint.json"
+    ]
+  },
+  "targetDefaults": {
+    "test:lib": {
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^public"],
+      "outputs": ["{projectRoot}/coverage"]
+    },
+    "test:eslint": {
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^public"]
+    },
+    "test:types": {
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^public"]
+    },
+    "build": {
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^public"],
+      "outputs": ["{projectRoot}/build", "{projectRoot}/dist"]
+    },
+    "test:build": {
+      "dependsOn": ["build"],
+      "inputs": ["^public"]
+    }
+  }
+}
