+++ package.json
@@ -1,22 +1,27 @@
 {
   "private": true,
   "repository": "https://github.com/tanstack/store.git",
+  "packageManager": "pnpm@8.5.1",
+  "type": "module",
   "scripts": {
-    "clean-dist": "pnpm -rc --parallel exec 'rm -rf build dist'",
-    "clean-all": "pnpm -rc --parallel exec 'rm -rf build dist node_modules'",
-    "test": "is-ci-cli test:ci test:dev",
-    "test:ci": "pnpm --parallel test",
-    "test:dev": "vitest",
-    "buildNx": "nx run-many --target=build --projects=@tanstack/* --exclude=@tanstack/react-start",
-    "build": "rollup --config rollup.config.js && pnpm typecheck",
-    "watch": "concurrently --no-color --prefix \"watch\" --kill-others --raw \"rollup --config rollup.config.js -w\" \"tsc -b --watch --preserveWatchOutput\"",
-    "dev": "pnpm watch",
-    "prettier": "prettier \"packages/*/{src/**,examples/**/src/**}.{md,js,jsx,ts,tsx,json}\" --write",
-    "visualize": "pnpm -rc --parallel exec open build/stats-html.html",
-    "cipublish": "ts-node scripts/publish.ts",
-    "cipublishforce": "CI=true pnpm cipublish",
-    "typecheck": "tsc -b",
-    "typecheck-perf": "tsc -b --generateTrace ts-perf"
+    "clean": "pnpm --filter \"./packages/**\" run clean",
+    "preinstall": "node -e \"if(process.env.CI == 'true') {console.log('Skipping preinstall...'); process.exit(1)}\" || npx -y only-allow pnpm",
+    "install:csb": "corepack enable && pnpm install --frozen-lockfile",
+    "test": "pnpm run test:ci",
+    "test:ci": "nx affected --exclude=examples/** --targets=test:format,test:eslint,test:lib,test:types,build,test:build",
+    "test:eslint": "nx affected --target=test:eslint",
+    "test:format": "pnpm run prettier --check",
+    "test:lib": "nx affected --target=test:lib",
+    "test:lib:dev": "pnpm --filter \"./packages/**\" run test:lib:dev",
+    "test:build": "nx affected --target=test:build",
+    "test:types": "nx affected --target=test:types",
+    "build": "nx affected --target=build",
+    "build:all": "nx run-many --exclude=examples/** --target=build",
+    "watch": "pnpm run build:all && nx watch --all -- pnpm run build:all",
+    "dev": "pnpm run watch",
+    "prettier": "prettier \"{packages,examples,scripts}/**/*.{md,js,jsx,cjs,ts,tsx,json,vue}\"",
+    "prettier:write": "pnpm run prettier --write",
+    "cipublish": "node scripts/publish.js"
   },
   "workspaces": [
     "./packages/*",
@@ -61,12 +66,13 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-virtual": "^2.10.4",
-    "rollup": "^2.70.2",
-    "rollup-plugin-dts": "^4.2.2",
+    "rollup": "^3.23.0",
+    "rollup-plugin-node-externals": "^6.1.0",
+    "rollup-plugin-preserve-directives": "^0.2.0",
+    "rollup-plugin-dts": "^6.0.0",
     "rollup-plugin-size": "^0.2.2",
-    "rollup-plugin-svelte": "^7.1.0",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-visualizer": "^5.6.0",
+    "rollup-plugin-visualizer": "^5.9.0",
     "solid-js": "^1.6.5",
     "stream-to-array": "^2.3.0",
     "svelte": "^3.55.0",
