--- rollup.config.js
@@ -1,7 +1,0 @@
-require('ts-node').register({
-  compilerOptions: {
-    esModuleInterop: true,
-  },
-})
-
-module.exports = require('./rollup.config.ts')
