+++ scripts/types.d.ts
@@ -1,3 +1,5 @@
+import type { RollupOptions } from 'rollup'
+
 export type Commit = {
   commit: CommitOrTree
   tree: CommitOrTree
@@ -20,7 +22,7 @@ export type AuthorOrCommitter = {
 }
 
 export type Parsed = {
-  type: string
+  type: string | null
   scope?: string | null
   subject: string
   merge?: null
@@ -37,22 +39,9 @@ export type Parsed = {
 export type Package = {
   name: string
   packageDir: string
-  srcDir: string
-  builds: Build[]
-}
-
-export type Build = {
-  jsName: string
-  entryFile: string
-  external?: (d: string) => any
-  globals?: Record<string, string>
-  esm?: boolean
-  cjs?: boolean
-  umd?: boolean
-  externals?: any[]
+  entries: Array<'main' | 'module' | 'svelte' | 'types'>
 }
 
 export type BranchConfig = {
   prerelease: boolean
-  ghRelease: boolean
 }
